disco: correctly reply to queries from the server
If the server discos us (as it will when checking our XEP-0115
capabilities), the stanza may well have no from='' attribute.
Previously, we would send a reply to the literal JID 'null', which of
course the server ignored.

With this patch, we omit to='' on the reply if the request omitted
from=''.